backend/requirements.txt
fastapi
uvicorn
requests
scikit-learn
numpy
Pillow
scipy